test_requirements.txt
django-nose
django-libs
coverage
django-coverage
ipdb
flake8
fabric
factory_boy
mock
